separate reusable fixtures for plugin testing
